perf: replace lru-cache module with a simpler map
The lru-cache module is nice, but does lots of things we don't need in
the tokenizer, so it isn't as fast as it could be. This commit replaces
it with our own simple implemntation that's a good bit fastr.

This reduces tokenization time by a further 19%:

![](https://memes.peet.io/img/24-04-bcdbe2e7-853b-401b-ba54-7aab957bdf64.png)

This builds on #40 and closes #37 -- only the last commit is new.

Note that as of this commit, the vast plurality of time (45%) is doing
the map lookup in the cache, to the point that removing the cache doesn't
actually seem to impact the tokenization runtime very much because byte
pair encoding is so fast! But it seems a good point of scrutiny for any
further speed ups.

cc @andreamah